[diem adapter] introduce VMAdapter trait
Closes: #8773
Closes: #10648
