+++ vm-validator/src/vm_validator.rs
@@ -10,7 +10,7 @@ use diem_types::{
     on_chain_config::{DiemVersion, OnChainConfigPayload, VMConfig, VMPublishingOption},
     transaction::{SignedTransaction, VMValidatorResult},
 };
-use diem_vm::DiemVMValidator;
+use diem_vm::DiemVM;
 use fail::fail_point;
 use std::{convert::TryFrom, sync::Arc};
 use storage_interface::{state_view::VerifiedStateView, DbReader};
@@ -32,7 +32,7 @@ pub trait TransactionValidation: Send + Sync + Clone {
 #[derive(Clone)]
 pub struct VMValidator {
     db_reader: Arc<dyn DbReader>,
-    vm: DiemVMValidator,
+    vm: DiemVM,
 }
 
 impl VMValidator {
@@ -47,13 +47,13 @@ impl VMValidator {
             &smt,
         );
 
-        let vm = DiemVMValidator::new(&state_view);
+        let vm = DiemVM::new_for_validation(&state_view);
         VMValidator { db_reader, vm }
     }
 }
 
 impl TransactionValidation for VMValidator {
-    type ValidationInstance = DiemVMValidator;
+    type ValidationInstance = DiemVM;
 
     fn validate_transaction(&self, txn: SignedTransaction) -> Result<VMValidatorResult> {
         fail_point!("vm_validator::validate_transaction", |_| {
@@ -86,7 +86,7 @@ impl TransactionValidation for VMValidator {
         let version = config.get::<DiemVersion>()?;
         let publishing_option = config.get::<VMPublishingOption>()?;
 
-        self.vm = DiemVMValidator::init_with_config(version, vm_config, publishing_option);
+        self.vm = DiemVM::init_with_config(version, vm_config, publishing_option);
         Ok(())
     }
 }
